requirements.txt
pipecat-ai[google,silero,webrtc]>=0.0.102
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiosqlite>=0.20.0
python-dateutil>=2.9.0
openai>=1.0.0
